Views do not notify count changes properly
The views don't have a valid count until both a model and a delegate
are provided.  But, countChanged() is not called when a delegate is
set after the model, so bindings to count were not evaluated.  Also
test that count is updated for insertion/removal

Change-Id: Ic82039a624c02f0bdb2862ac7a6e215df75bb314
Task-number: QTBUG-19037
Reviewed-by: Bea Lam
(cherry picked from commit 4fbcb6a531bf424a8dbbbd0ca2947d9e15118885)