Fix potential crash when displaying multiscripted text
Shaping has to be done in the current thread, otherwise the
font engines index for each glyph (referenced in the msb of the glyph
index) might not be valid yet, because the font engines list is
populated when shaping is done. So we need to make sure that the
render thread relayouts. Geometry changes will already cause a
relayout, which will in turn cause another relayout when the
paint node is updated. There doesn't seem to be any convenient
and safe way of avoiding this doubling of the layout step if we
want to have rendering in a different thread than the QML graph.

Reviewed-by: Gunnar