Add VisualListModel
The VisualListModel is a variant of VisualDataModel that looks like
a list model with add, remove etc. methods but it stores QSGItems
rather than object data. The aim is to provide a visual model of a
list to allow dragging/dropping and general editing of a "visual model"
without affecting the underlying list model for a view. This allows
a view to apply visual changes which may or may not be later be
synchronized with the underlying list model.