QmlInspector: Removed private header postfix and Qt namespace
Renamed the headers back to normal since they are not included in a Qt
library. Also took the classes out of the Qt namespace and back into
the QmlJSDebugger namespace.

This is mainly to make it easier again to port changes back into the
version of the inspector shipped with Qt Creator.

Change-Id: I820c2372089c24a42404b4055be1b938e384e352