Performance improvements
There are two main changes here.  First, where possible, we mark
properties as "IsDirect" which means that they exist in a C++
QMetaObject (as opposed to a dynamic meta object), which allows us
to call QObject::qt_metacall() directly, bypassing any dynamic meta
object stuff.

The second change is to use an ascii string comparator in V8 where
possible.  V8 stores ASCII string internally as ASCII strings, and
asking it to compare them to a UTF16 string requires a conversion.