qmlplugindump: Fix dumping with -path on Mac.
If the current working directory was a direct parent of the qmldir path
the exported modules had the path as the module URI on macs.

Also changes the QtQuick export back to 1.0 to make it work with
Qt 4.7.3. The version of that import statement does not actually matter
as long as it's valid.

Mirrors a change in Qt Creator:
http://codereview.qt.nokia.com/896

(cherry picked from commit d7ab0007d4b051f3cf12f01157b8b78d2fddf7c8)

Change-Id: Ib8caa8eda7f237b10e82df3afd560c118427aa8f
Reviewed-on: http://codereview.qt.nokia.com/897
Reviewed-by: Kai Koehne <kai.koehne@nokia.com>