Add support for a vector4d type in QML
QVector4D is a value-type which is supported but was not able to be
constructed using a Qt object function.  This commit allows properties
of vector4d type to be constructed, and adds a function to the global
Qt object and adds unit tests to ensure that it behaves correctly.

Task-number: QTBUG-18559
Change-Id: I96509a4f496b644d20fdb1d977d0afe430d89e13
Reviewed-on: http://codereview.qt.nokia.com/1626
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Aaron Kennedy <aaron.kennedy@nokia.com>