Don't clip the cursor at the far right of a TextInput.
The cursor is allowed to exceed the boundaries of a TextInput so that
the text alignment is consistent with the Text element and the bounding
rect of TextInput is expanded to allow for that.  To avoid clipping
the contentSize also needs to be expanded, and the horizontal scrolling
should be calculated against the non-expanded size.

Task-number: QTBUG-18818
Change-Id: I09e3a7aff9c2e8a333e8a3f9796683563b3e124a
Reviewed-on: http://codereview.qt.nokia.com/2242
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Martin Jones <martin.jones@nokia.com>