Renamed ShaderEffectItem to ShaderEffect and some API changes.
The ShaderEffectItem was renamed to ShaderEffect to be consistent
with other QML element names. The GLSL uniform variable
qt_ModelViewProjectionMatrix was renamed to qt_Matrix which is
easier to type and remember. The GridMesh element was removed.
The mesh resolution is now specified simply with a QSize.
To make the transition easier, the old API will continue to work
for some time, but will output warnings if used. Eventually, I
will remove the old API completely.

Change-Id: Iec4f2aa624a2c76a7db6750c58f73dbcb316ab6a
Reviewed-on: http://codereview.qt.nokia.com/2270
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Gunnar Sletta <gunnar.sletta@nokia.com>