Fix content position for key navigation with StrictlyEnforceRange
If a ListView had highlight ranges and StrictlyEnforceRange, but no
highlight item, the content would not move to the correct position
when incrementCurrentIndex() and decrementCurrentIndex() were invoked.

trackedPositionChanged() shouldn't take the current section pos into
account because this is already calculated by FxListItemSG::position()
(this wasn't the case when the code in trackedPositionChanged() was
originally written).

Task-number: QTBUG-20287
Change-Id: I1624b5afd1efbe27630349143b7af2b486cfa260
Reviewed-on: http://codereview.qt.nokia.com/2429
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Bea Lam <bea.lam@nokia.com>
Reviewed-by: Martin Jones <martin.jones@nokia.com>