Make QColor a value type
This commit allows direct access to the r, g, b and a components
of a color (in floating point format: 0 <= v <= 1). Since conversion
from color to string is a common operation, this commit also adds
unit tests to ensure that the previous behaviour is maintained in
other cases (comparison with toString value, etc).

Task-number: QTBUG-14731
Change-Id: I87b521dd4f9c1e96dfe5b20cf8053293cb14cfe4
Reviewed-on: http://codereview.qt.nokia.com/2527
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Aaron Kennedy <aaron.kennedy@nokia.com>