Fix ListView refill() where delegate size == 0
When refilling, if a delegate item is positioned at the start of the
buffer, it should only be removed if its height is > 0. Otherwise
refill() will constantly remove it in removeNonVisibleItems() and
add it again in the next frame in addVisibleItems().

Regression from aa99d4f5cb4f1f512d1db90536752b518dbea449

Task-number: QTBUG-20712
Change-Id: I21d25fde21e0b559a378c9e202ace58947417a4d
Reviewed-on: http://codereview.qt.nokia.com/2541
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Martin Jones <martin.jones@nokia.com>