Add attached properties to the positioners.
Added an attached property type to the base positioner class.
This allows items within a positioner (Row, Column, Grid, Flow) to
determine their index inside the positioner, as well as whether
they are the first or last items. Non-visible items are ignored,
as in the positioner layout objects themselves. It may be useful
to expand this in the future to contain more information specific
to the positioner, for example row/column for Grid.

Task-number: QTBUG-19211
Change-Id: I10a8c9ca5528dd12811125cae8a9b4d2e3747972
Reviewed-on: http://codereview.qt.nokia.com/2983
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Alan Alpert <alan.alpert@nokia.com>