Allow initial property values to be defined with QSGLoader
This commit adds an "active" property to QSGLoader, which can be
used to delay instantiation of the item until the user wishes to
activate the loader.  The property is true by default in order to
maintain compatibility with previous behaviour.

The commit also adds a "setSource(v8object)" function to QSGLoader,
which behaves identically to setSource() property mutator except that
it takes a JavaScript object parameter which defines the initial
property values of the item (in a manner similar to that of
QDeclarativeComponent::createObject()).

Task-number: QTBUG-17009
Change-Id: Ifd824b518b60ef7aa3017c384835abb552e65cf1
Reviewed-on: http://codereview.qt.nokia.com/3364
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Michael Brasser <michael.brasser@nokia.com>