Reintroduce the QSGTextureProvider as a separate object
The texture provider needs to live completely in the rendering thread
and needs to be selfcontained during rendering. During the items
updatePaintNode() all needed state is copied into the texture
provider. Texture providers have thread affinity in the rendering
thread and are deleted there using deleteLater() and the added
processEvents() in the rendering loop.

This fixes the bug where a QSGItem is deleted as a result from an
animation while the same item is being used as a texture provider in
the rendering thread.

There is also a small optimzation to QSGShaderEffectSource
in that we don't create a paint node when the shader source is
not going to be shown.

Change-Id: I6b9bc1da2a0f55d3d5356d4091fd6af6a7ea6f01
Reviewed-on: http://codereview.qt.nokia.com/3293
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Kim M. Kalland <kim.kalland@nokia.com>