Do not deliver mouse wheel events when item is not visible
This patch change the code to also skip invisible items (it was
done for disabled already) when deciding whether or not deliver the
wheel event to an item. The rationale here is to follow the same
rule as the other mouse events.

An autotest was added to verify and maintain this behavior.

Change-Id: If0fe6d64d1f7cfb8679ce11edda7c02dc3783f94
Reviewed-on: http://codereview.qt.nokia.com/3429
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Michael Brasser <michael.brasser@nokia.com>