Added --verify-only option to qmlmin.
qmlmin has three different stages. In the first stage
it generates the QML/JS minified code. In the second
stage we verify that minified code is equivalent
to the original code and in the final stage we
produce the output. With --verify-only you can tell
qmlmin to quit after the verification step.

Note that this option is pretty much equivalent to
the unix command qmlmin file.qml -o /dev/null.

Change-Id: I91373bc1c1db8c35af2e301ad13d7b34fc384529
Reviewed-on: http://codereview.qt-project.org/4670
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Kent Hansen <kent.hansen@nokia.com>