Fixed compile on mac (workaround for build system bug)
When using QT += somemodule, and using frameworks on mac, qmake refuses
to add `-framework somemodule' to the compiler flags unless the
framework can be found on disk.  This can easily break when compiling
two frameworks out of the same source tree.

In this case, if QtQuick1 were qmake'd prior to QtDeclarative being
built (which is expected for a clean build), it would incorrectly put
-lQtDeclarative into the link line even if QtDeclarative should be
built as a framework.  The problem would disappear if qmake was re-run.

Change-Id: I79cbfc454e0ab564ce6597b4f7e613c1ff39473f
Reviewed-on: http://codereview.qt-project.org/4938
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Michael Brasser <michael.brasser@nokia.com>