Support preedit text in QSGTextInput
We need to support merged additional formats with backgrounds
in the QSGTextInput as well, so that code has been separated
into its own function. We also need to account for the position
in the bounding rect, so that the decorations are painted at
the correct location when there text input is scrolled.

Task-number: QTBUG-21261
Change-Id: I0799a62bf26e6a7a2c1a6eef9bbdf889c1c8e870
Reviewed-on: http://codereview.qt-project.org/4964
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Jiang Jiang <jiang.jiang@nokia.com>