Get rid of redundant type member from V4 instructions
When interpretation is threaded, the type member is
redundant, since we can obtain the type from the
instruction address.

Getting rid of the type member can save at least one
byte per instruction (possibly more because of
alignment). On ia32, the Noop and Block instructions
become 4 bytes smaller.

The compiler has been refactored to leave it up to
the Bytecode class to decide whether the instruction
address or type (enum) should be stored.

To achieve this, the "low-level" Instr union was
renamed to V4Instr, and the new Instr struct
used by the compiler provides typedefs for
instantiating the data used by each particular
instruction. Lastly, the gen() function (instruction
emitter) was made template-based to provide the
instruction type.

The instructions Bool, Int, Real, String were
renamed to Load{Bool,Int,Real,String} to make the
new code more readable.

This approach follows a similar refactoring in the
QML-VME interpreter.

Change-Id: I9bf16b099a85afdfb719a25fb18047408882d61e
Reviewed-on: http://codereview.qt-project.org/5393
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Roberto Raggi <roberto.raggi@nokia.com>