Allow view items to be preserved.
Do not destroy items that are members of the VisualDataModel
persistedItems group when they are released by the view.

The create function on VisualDataGroup will return a reference to
an instantiated item within that group, any item returned by this
function is automatically added to the persistedItems group and can
be released by removing it from the group.  Uninstantiated items added
to the persistedItems group by any other means are not instantiated
until requested by a view or create.

Task-number: QTBUG-21518
Task-number: QTBUG-20854

Change-Id: I59554711208504c8f20a3ebe783bddab9b21a558
Reviewed-on: http://codereview.qt-project.org/5831
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Martin Jones <martin.jones@nokia.com>