(private-) export of QSGShaderEffectSource.
QtLocation maps need to be able to render standard QML elements as
Map items (for example a convenience store logo and text). Prior to
Qt5 this was achieved with Qt graphics view framework.

Only really private symbol used, in addition to inheriting a class from
QSGShaderEffectSource, is textureProvider(). It is used to access
texture IDs, which can then be used to map textures to map item
geometries.

Change-Id: Ifc6cf1d20b11493738c77b3bde9972a8cb1c70ce
Reviewed-on: http://codereview.qt-project.org/5912
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Gunnar Sletta <gunnar.sletta@nokia.com>