qmlplugindump: Add test for running with -builtins.
The main idea is to check for crashes: When instantiable C++ based
QML components have uninitialized QObject * members, qmlplugindump can
crash while collecting types to dump. That will make it impossible
for users to use qmlplugindump and should be fixed in the responsible
classes constructor.

On unix it has a crash handler that tries to pinpoint which class
and property caused the problem.

Change-Id: Ie049cae27d94920f3d8346ca7075eba5a11c2a37
Reviewed-on: http://codereview.qt-project.org/5976
Reviewed-by: Roberto Raggi <roberto.raggi@nokia.com>