Fix floating images in QSGTextNode
When images are floating and we encounter the object replacement
character, we should ignore them, as they're done by special
handling of the QTextFrame. We also need to make sure they are
not attempted inlined by passing the layout position of the frame
to addTextObject(), and we need to set a valid QTextLine for the
object to make sure it's added to the graph.

Change-Id: Ia43ef3fd4bb9c7b48e0862071118ee0f176212bd
Reviewed-on: http://codereview.qt-project.org/6036
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Jiang Jiang <jiang.jiang@nokia.com>