Debugger: Fix lock when profiling
Parts of the loading of data is now in a separate thread, which was already
trying to log trace data while the QDeclarativeDebugTrace constructor
was still blocking. Avoid this by calling QDeclarativeDebugTrace constructor
before the actual execution begins (addEngine()). Also make sure that the
logging methods are reentrant with a mutex.

Change-Id: I5c2c1d14763fd9c7cb6fc93c6dff22d00d8737f1
Reviewed-on: http://codereview.qt-project.org/6169
Reviewed-by: Christiaan Janssen <christiaan.janssen@nokia.com>
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>