Stabilize Particles tests
One potential problem is that in certain circumstances the recycling
data list is so efficient that it doesn't need to use all particle data
in the list. In those cases, the uninitialized members were causing test
failures but these slight differences are acceptable for the simulation.

The ParticleSystem test is extended to check that it's only small edge
cases.

Change-Id: I5e2f133dc4b96dd05d49e7e983ae630e379fa66a
Reviewed-on: http://codereview.qt-project.org/6312
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Martin Jones <martin.jones@nokia.com>