Remove the InputContext module.
The input context module provided a bridge between QML and a
QInputContext which allowed an application to implement it's own input
method in-process.  However since QInputContext has been replaced and
now only exists as a QWidget compatibility API this module no longer
works with the QtQuick 2 items and porting it to the new
QPlatformInputContext API does not seem to be an option as there is
no way to override the instance provided by the platform integration.

Change-Id: Ica602377436fcea42b9ffd05e1b347e2ee52acc8
Reviewed-by: Martin Jones <martin.jones@nokia.com>