(private-) export QSGShaderEffectTexture
QtLocation maps need to be able to render standard QML elements
as Map items (for example a convenience store logo and text).
Prior to Qt5 this was achieved with Qt graphics view framework.

This export is needed in order to be able to ask QML engine
to generate/update/render textures as needed.
Only private symbol really used is the updateTextures().

This commit complements the related earlier commit
1780033cb259bbb166fe9b4b14c92829a8e1800e

Change-Id: I98b2083b6a4d25fdd113ac4d95a9b9a2bb13a1f1
Reviewed-on: http://codereview.qt-project.org/6475
Reviewed-by: Lars Knoll <lars.knoll@nokia.com>
Reviewed-by: Samuel Rødal <samuel.rodal@nokia.com>