Fix incorrect cache indexes when consecutive list items are removed.
When cached model items are removed listItemsRemoved will try and append
a new cache only range onto a preceding one where possible. Previously
it would then re-run that range so the current indexes in the list were
incremented before moving onto the next range, which meant that the
current indexes were incremented by size of the previous range twice.

Instead of processing the range a second time increment the cache
indexes directly when inserting a cache only range and move directly
onto the next range in the list.

Change-Id: I63418c4397f911cefb521c5a5b0dd25faf66e08b
Reviewed-by: Martin Jones <martin.jones@nokia.com>