Fold Move into Affector
Since 'Move' was just a generic physics property affector, it makes
sense to fold it into the generic affector as an easier way of
implementing the common case of custom affector (and much faster, since
it doesn't hit JS).

For completeness in the Move behavior, 'once' now applies the effect of
affectors over a second in that one time, so as to avoid exposing the
simulation tick. This allows once to have actual meaning for continous
affectors.

Gravity is also going to be kept around as a convenience (although it
will change slightly), so just removing its deprecation warning.

Change-Id: Ieb52f8990445fe95f94070175a0f9beb6686324b
Reviewed-by: Martin Jones <martin.jones@nokia.com>