Force V8 debugger load on startup instead of on demand.
The V8 debugger code can take several hundred milliseconds to
start. This is by default loaded and unloaded on demand, for
example whenever a script is compiled. This has a significant
effect on the timing reported by the QML profiler in Qt
Creator. This change forces the debugger context to be loaded
when the v8 context is created, so that this cost is paid
once on startup and no longer affects the QML profiling times.

Task-number: QTBUG-22170
Change-Id: I5533b53908b3c53e18b0eb870ed6166d109429df
Reviewed-by: Aaron Kennedy <aaron.kennedy@nokia.com>