Fix bus errors on Linux/ARM due to word alignment issues
If the kernel is configured to SIGBUS on unaligned word access, then the code
in qdeclarativecompiler.cpp that accesses the structs that follow
QDeclarativeVMEMetaData in memory may cause the process to get killed. They
contain integers but they're not aligned.

Ensure that the size of the first structure gives integer alignment for the
following structures.

Change-Id: Iad85557aa3be4949f653d389dbaa3666960be404
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>