Optimization: Minimize refreshing when root context changes
Modifying a context requires expressions to be refreshed incase they
depend on a name resolution that changes.  As it is common to modify
the root context, and the root context can't hide any names, it is easy
to constrain the expressions that require refreshing.  In a well
behaving app, this should mean that no expressions are reevaluated when
the root context is modified.

Change-Id: Id3b48cd595fdd6b5a3dc6f26319f652dccaef79c
Reviewed-by: Roberto Raggi <roberto.raggi@nokia.com>