Support JS objects in ListModel via QVariantMap
Added support for the new listmodel implementation to store and
retrieve JS objects via QVariantMap. Storing JS objects in a
listmodel is significantly slower than storing native datatypes
at the moment (this may be improved in the future). Also note
that it's not currently possible to bind to fields within the JS
object.

Change-Id: I3b1a11ace7cdec754c1a2bb2b2d1b7edf561864d
Reviewed-by: Martin Jones <martin.jones@nokia.com>