qml debugger: Create abstraction for State management
State and PropertyChanges aren't part of the core qml types. The
(general-purpose) qml debugger shouldn't have to know those types;
they are only relevant when QtQuick (2.0) has been imported.

Introduce a delegate interface that performs State processing.
Implement a delegate for QtQuick 2, and install it when the QtQuick 2
module is imported.

Change-Id: I8af1157346ca9365eb0f7b99ccb71744d17fcebc
Reviewed-by: Aurindam Jana <aurindam.jana@nokia.com>